paper2demo/requirements.txt
PyMuPDF>=1.24.0
google-generativeai>=0.7.0
openai>=1.30.0
Pillow>=10.0.0
